+++ src/nsidc/metgen/netcdf_reader.py
@@ -3,7 +3,6 @@ Interface functions that read various metadata attribute values
 from source science data files.
 """
 
-import json
 import logging
 import os.path
 import re
@@ -34,7 +33,7 @@ def extract_metadata(netcdf_path, configuration):
         "size_in_bytes": os.path.getsize(netcdf_path),
         "production_date_time": date_modified(netcdf, configuration),
         "temporal": time_range(os.path.basename(netcdf_path), netcdf, configuration),
-        "geometry": {"points": json.dumps(spatial_values(netcdf, configuration))},
+        "geometry": {"points": spatial_values(netcdf, configuration)},
     }